refactor(pipeline): move logger to shared module
Extract PipelineLogger interface and createLogger factory from stages/types.ts into shared/logger.ts. This establishes a shared infrastructure location separate from domain-specific stage types, improving separation of concerns and module organization.
