+++ apps/pipeline/src/cli.ts
@@ -4,10 +4,10 @@ import { showStatus } from '@/commands/status.ts';
 import { EXIT_CODES, PATHS } from '@/config.ts';
 import type { PipelineOptions } from '@/pipeline.ts';
 import { PipelineError, runPipeline } from '@/pipeline.ts';
+import type { PipelineLogger } from '@/shared/logger.ts';
+import { createLogger } from '@/shared/logger.ts';
 import { syncDescriptions } from '@/stages/sync-descriptions.ts';
 import { syncEraSummaries } from '@/stages/sync-era-summaries.ts';
-import type { PipelineLogger } from '@/stages/types.ts';
-import { createLogger } from '@/stages/types.ts';
 import { validateAllDescriptions } from '@/stages/validate-descriptions.ts';
 
 function parseYearOption(raw: string): number {
