+++ apps/pipeline/src/commands/list-years.ts
@@ -1,6 +1,6 @@
 import { PATHS } from '@/config.ts';
+import type { PipelineLogger } from '@/shared/logger.ts';
 import { parseYearsFromDirectory } from '@/stages/fetch.ts';
-import type { PipelineLogger } from '@/stages/types.ts';
 
 export async function listYears(logger: PipelineLogger): Promise<void> {
   const years = await parseYearsFromDirectory(PATHS.sourceGeojson);
