+++ apps/pipeline/src/shared/logger.ts
