+++ apps/pipeline/src/stages/convert.test.ts
@@ -5,8 +5,8 @@ vi.mock('@/exec.ts', () => ({
 }));
 
 import { execFileAsync } from '@/exec.ts';
+import type { PipelineLogger } from '@/shared/logger.ts';
 import { buildTippecanoeArgs, executeConvert } from '@/stages/convert.ts';
-import type { PipelineLogger } from '@/stages/types.ts';
 
 const mockExecFileAsync = vi.mocked(execFileAsync);
