+++ apps/pipeline/src/stages/index-gen.test.ts
@@ -2,8 +2,8 @@ import { mkdtemp, rm, writeFile } from 'node:fs/promises';
 import { tmpdir } from 'node:os';
 import path from 'node:path';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
+import type { PipelineLogger } from '@/shared/logger.ts';
 import { generateYearIndex } from '@/stages/index-gen.ts';
-import type { PipelineLogger } from '@/stages/types.ts';
 
 function createMockLogger(): PipelineLogger {
   return {
