+++ apps/pipeline/src/stages/sync-descriptions.ts
@@ -2,7 +2,7 @@ import { mkdirSync, writeFileSync } from 'node:fs';
 import path from 'node:path';
 import { Client } from '@notionhq/client';
 import type { PageObjectResponse } from '@notionhq/client/build/src/api-endpoints';
-import type { PipelineLogger } from '@/stages/types.ts';
+import type { PipelineLogger } from '@/shared/logger.ts';
 import { validateDescriptionFile } from '@/stages/validate-descriptions.ts';
 
 interface TerritoryProfile {
