+++ apps/pipeline/src/stages/validate.test.ts
@@ -3,7 +3,7 @@ import { mkdtemp, rm } from 'node:fs/promises';
 import { tmpdir } from 'node:os';
 import path from 'node:path';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
-import type { PipelineLogger } from '@/stages/types.ts';
+import type { PipelineLogger } from '@/shared/logger.ts';
 import { runValidateForYear } from '@/stages/validate.ts';
 
 const FIXTURES = path.join(process.cwd(), 'tests', 'fixtures');
